+++ src/main.rs
@@ -82,7 +82,9 @@ mod tests {
         assert!(output.status.success());
 
         // Check if output files are created
-        assert!(std::path::Path::new(&format!("{}/reproductive_fitness_0_table.npy", tmp.dir)).exists());
+        assert!(
+            std::path::Path::new(&format!("{}/reproductive_fitness_0_table.npy", tmp.dir)).exists()
+        );
         assert!(std::path::Path::new(&format!("{}/virolution.log", tmp.dir)).exists());
     }
 
@@ -108,8 +110,12 @@ mod tests {
         assert!(output.status.success());
 
         // Check if output files are created
-        assert!(std::path::Path::new(&format!("{}/reproductive_fitness_0_table.npy", tmp.dir)).exists());
-        assert!(std::path::Path::new(&format!("{}/reproductive_fitness_1_table.npy", tmp.dir)).exists());
+        assert!(
+            std::path::Path::new(&format!("{}/reproductive_fitness_0_table.npy", tmp.dir)).exists()
+        );
+        assert!(
+            std::path::Path::new(&format!("{}/reproductive_fitness_1_table.npy", tmp.dir)).exists()
+        );
         assert!(std::path::Path::new(&format!("{}/virolution.log", tmp.dir)).exists());
     }
 
@@ -135,8 +141,14 @@ mod tests {
         assert!(output.status.success());
 
         // Check if output files are created
-        assert!(std::path::Path::new(&format!("{}/reproductive_fitness_0_table.npy", tmp.dir)).exists());
-        assert!(std::path::Path::new(&format!("{}/reproductive_fitness_0_epistasis_table.npy", tmp.dir)).exists());
+        assert!(
+            std::path::Path::new(&format!("{}/reproductive_fitness_0_table.npy", tmp.dir)).exists()
+        );
+        assert!(std::path::Path::new(&format!(
+            "{}/reproductive_fitness_0_epistasis_table.npy",
+            tmp.dir
+        ))
+        .exists());
         assert!(std::path::Path::new(&format!("{}/virolution.log", tmp.dir)).exists());
     }
 }